Added Darlinton support, Calibration cleanups, PROGMEM icon fixed
Added a subtype field to the checks struct, and added subtypes 'TYPE_REGULAR' and 'TYPE_DARLINGTON' for BJTs. Added code to save and display this information.

Made the Calibration functions much neater. Fixed up button presses when in Calibration mode.

Fixed the generation of custom character icons from Flash, which was borked.